Changes in logic

+++ iPullTableToZoom/ViewController.m
@@ -47,7 +47,7 @@
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:@"cell" forIndexPath:indexPath];
-    [cell.textLabel setText:[NSString stringWithFormat:@"Row - %d",indexPath.row]];
+    [cell.textLabel setText:[NSString stringWithFormat:@"Row - %ld",(long)indexPath.row]];
     return cell;
 }
